backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
strands-agents==0.1.6
boto3==1.35.86
gitpython==3.1.43
